Added Profile for Peter Kirby
This is the only way I found to commit this new file. Tried with the App but no permissions. I also had prepared a .png for Peter's Headshot but again couldn't upload it. So the pic is an imgur link which can be downloaded and added as peter-kirby.png at your end. Do you have an email I can use to have a chat on the next projects/profiles to create?